move deployed future into its own module
